+++ apps/jan-api-gateway/application/app/domain/auth/auth_service.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/golang-jwt/jwt/v5"
-	"menlo.ai/jan-api-gateway/app/domain/apikey"
 
+	"menlo.ai/jan-api-gateway/app/domain/apikey"
 	"menlo.ai/jan-api-gateway/app/domain/organization"
 	"menlo.ai/jan-api-gateway/app/domain/user"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/requests"
@@ -233,4 +233,74 @@ func GetUserIDFromContext(reqCtx *gin.Context) (string, bool) {
 
 func SetUserIDToContext(reqCtx *gin.Context, v string) {
 	reqCtx.Set(string(UserContextKeyID), v)
+}
+
+type ApikeyContextKey string
+
+const (
+	ApikeyContextKeyEntity   ApikeyContextKey = "ApikeyContextKeyEntity"
+	ApikeyContextKeyPublicID ApikeyContextKey = "apikey_public_id"
+)
+
+func (s *AuthService) GetAdminApiKeyFromQuery() gin.HandlerFunc {
+	return func(reqCtx *gin.Context) {
+		ctx := reqCtx.Request.Context()
+		user, ok := GetUserFromContext(reqCtx)
+		if !ok {
+			reqCtx.AbortWithStatusJSON(http.StatusUnauthorized, responses.ErrorResponse{
+				Code: "72ca928d-bd8b-44f8-af70-1a9e33b58295",
+			})
+			return
+		}
+
+		publicID := reqCtx.Param(string(ApikeyContextKeyPublicID))
+		if publicID == "" {
+			reqCtx.AbortWithStatusJSON(http.StatusBadRequest, responses.ErrorResponse{
+				Code:  "9c6ed28c-1dab-4fab-945a-f0efa2dec1eb",
+				Error: "missing apikey public ID",
+			})
+			return
+		}
+
+		adminKeyEntity, err := s.apiKeyService.FindOneByFilter(ctx, apikey.ApiKeyFilter{
+			PublicID: &publicID,
+		})
+		if adminKeyEntity == nil || err != nil {
+			reqCtx.AbortWithStatusJSON(http.StatusNotFound, responses.ErrorResponse{
+				Code: "f4f47443-0c80-4c7a-bedc-ac30ec49f494",
+			})
+			return
+		}
+		memberEntity, err := s.organizationService.FindOneMemberByFilter(ctx, organization.OrganizationMemberFilter{
+			UserID:         &user.ID,
+			OrganizationID: adminKeyEntity.OrganizationID,
+		})
+		if memberEntity == nil || err != nil {
+			reqCtx.AbortWithStatusJSON(http.StatusUnauthorized, responses.ErrorResponse{
+				Code: "56a9fa87-ddd7-40b7-b2d6-94ae41a600f8",
+			})
+			return
+		}
+		SetAdminKeyFromContext(reqCtx, adminKeyEntity)
+	}
+}
+
+func GetAdminKeyFromContext(reqCtx *gin.Context) (*apikey.ApiKey, bool) {
+	apiKey, ok := reqCtx.Get(string(ApikeyContextKeyEntity))
+	if !ok {
+		return nil, false
+	}
+	v, ok := apiKey.(*apikey.ApiKey)
+	if !ok {
+		return nil, false
+	}
+	return v, true
+}
+
+func SetAdminKeyFromContext(reqCtx *gin.Context, apiKey *apikey.ApiKey) {
+	reqCtx.Set(string(ApikeyContextKeyEntity), apiKey)
+}
+
+func AdminOrganizationMiddleware() gin.HandlerFunc {
+	
 }
