+++ apps/jan-api-gateway/application/cmd/server/wire_gen.go
@@ -52,14 +52,14 @@ func CreateApplication() (*Application, error) {
 	projectRepository := projectrepo.NewProjectGormRepository(transactionDatabase)
 	projectService := project.NewService(projectRepository)
 	organizationService := organization.NewService(organizationRepository, projectService)
-	apiKeyRepository := apikeyrepo.NewApiKeyGormRepository(transactionDatabase)
-	apiKeyService := apikey.NewService(apiKeyRepository)
 	userRepository := userrepo.NewUserGormRepository(transactionDatabase)
 	userService := user.NewService(userRepository)
-	adminApiKeyAPI := organization2.NewAdminApiKeyAPI(organizationService, apiKeyService, userService)
+	apiKeyRepository := apikeyrepo.NewApiKeyGormRepository(transactionDatabase)
+	apiKeyService := apikey.NewService(apiKeyRepository, organizationService)
+	authService := auth.NewAuthService(userService, apiKeyService, organizationService)
+	adminApiKeyAPI := organization2.NewAdminApiKeyAPI(organizationService, authService, apiKeyService, userService)
 	projectApiKeyRoute := apikeys.NewProjectApiKeyRoute(organizationService, projectService, apiKeyService, userService)
 	projectsRoute := projects.NewProjectsRoute(projectService, apiKeyService, projectApiKeyRoute)
-	authService := auth.NewAuthService(userService, apiKeyService, organizationService)
 	organizationRoute := organization2.NewOrganizationRoute(adminApiKeyAPI, projectsRoute, authService)
 	completionAPI := chat.NewCompletionAPI(apiKeyService)
 	chatRoute := chat.NewChatRoute(completionAPI)
